pyproject.toml
[tool.poetry]
name = "dry2-cli"
version = "1.0.0"
description = "PaaS-style CLI for DRY2 Infrastructure as Code"
authors = ["DRY2-IaaS Team"]
readme = "README.md"
packages = [{include = "dry2"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.0"
rich = "^13.0.0"
questionary = "^2.0.0"
pyyaml = "^6.0"
jinja2 = "^3.1.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
gitpython = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
dry2 = "dry2.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

